+++ NEWS.md
@@ -1,4 +1,4 @@
-# version 1.7-61
+# version 1.7-62
 
 ## bug fixes
 
@@ -20,8 +20,9 @@ https://stackoverflow.com/questions/77304534/rspatial-terraflip-error-when-flipp
 ## new
 
 - `k_means<SpatRaster>` method [#1314](https://github.com/rspatial/terra/issues/1314) by Agustin Lobo
+- `princomp<SpatRaster>` method [#1361](https://github.com/rspatial/terra/issues/1361) by Alex Ilich
 - `has.time<SpatRaster>` method 
-- new argument "raw=FALSE" to `rast`, `sds`, and `sprc` to ignore scale and offset [1354](https://github.com/rspatial/terra/issues/1354) by Insang Song
+- new argument "raw=FALSE" to `rast`, `sds`, and `sprc` to allow ignoring scale and offset [1354](https://github.com/rspatial/terra/issues/1354) by Insang Song
 
 
 # version 1.7-55